Add files via upload
The main.go file contains four different functions that provide users with options for searching, editing, and adding plants. Additionally, if the user wants to plant vegetables in their garden, they can type the name of the vegetable and the function will provide details including the name, PH level, water level, and soil type.

There are a few modifications that we need to make, but they can be implemented as we progress and add more functionality.